scripts/train_imitation.py
#!/usr/bin/env python3
"""
PPO + Imitation Learning training script for G1 robot.

This script implements PPO training with imitation learning components
for the Unitree G1 humanoid robot.
"""

import argparse
import logging
import os
from pathlib import Path

import hydra
from omegaconf import DictConfig, OmegaConf
import torch

# NOTE: These imports will need to be implemented
# from sonic_g1.env.g1_env import G1MotionImitationEnv
# from sonic_g1.models.policy import G1ActorCritic
# from sonic_g1.train.ppo_trainer import PPOTrainer
# from sonic_g1.data.dataset import ExpertDataset
# from sonic_g1.utils.config import load_config
# from sonic_g1.utils.logger import setup_logging

logger = logging.getLogger(__name__)


def setup_logging(level: str = "INFO"):
    """Setup logging configuration."""
    logging.basicConfig(
        level=getattr(logging, level.upper()),
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
        handlers=[
            logging.StreamHandler(),
            logging.FileHandler('train_imitation.log')
        ]
    )


@hydra.main(config_path="../configs/train", config_name="ppo_imitation", version_base="1.1")
def main(cfg: DictConfig) -> None:
    """Main training function."""

    # Setup
    setup_logging(cfg.get('log_level', 'INFO'))
    logger.info("Starting PPO + Imitation Learning training")
    logger.info(f"Config:\n{OmegaConf.to_yaml(cfg)}")

    # Set device
    if cfg.train.device == "auto":
        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    else:
        device = torch.device(cfg.train.device)

    logger.info(f"Using device: {device}")

    # TODO: Implement the following components
    logger.warning("Training components not yet implemented:")
    logger.warning("- G1MotionImitationEnv: MuJoCo environment for G1")
    logger.warning("- G1ActorCritic: Actor-critic policy architecture")
    logger.warning("- PPOTrainer: PPO training with imitation learning")
    logger.warning("- ExpertDataset: Expert demonstration loading")
    logger.warning("- Vectorized environment support")

    # Placeholder for actual implementation
    logger.info("This is a placeholder script. Implementation required:")
    logger.info("1. Create vectorized G1 environments")
    logger.info("2. Load expert demonstration data")
    logger.info("3. Create actor-critic policy model")
    logger.info("4. Setup PPO trainer with imitation loss")
    logger.info("5. Run training loop with periodic evaluation")
    logger.info("6. Save trained model and statistics")

    # Example structure (to be implemented):
    """
    # Create environments
    env_fn = lambda: G1MotionImitationEnv(cfg.env)
    envs = VectorEnv([env_fn] * cfg.train.ppo.num_envs)

    # Load expert data
    expert_dataset = ExpertDataset(cfg.train.imitation.expert_data_path, cfg)

    # Create model
    policy = G1ActorCritic(
        obs_dim=envs.observation_space.shape[0],
        action_dim=envs.action_space.shape[0],
        **cfg.train.model
    ).to(device)

    # Setup trainer
    trainer = PPOTrainer(
        policy=policy,
        envs=envs,
        expert_dataset=expert_dataset,
        cfg=cfg.train,
        device=device
    )

    # Train
    trainer.train()

    # Save final model
    torch.save({
        'policy_state_dict': policy.state_dict(),
        'config': cfg,
        'training_stats': trainer.get_stats()
    }, 'checkpoints/ppo_imitation/final_model.pt')
    """

    logger.info("Training completed (placeholder)")


if __name__ == "__main__":
    main()