pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "sonic-g1"
version = "0.1.0"
description = "SONIC G1: Unitree G1 Whole-Body Controller - MuJoCo-based training and deployment framework"
authors = [
    {name = "SONIC G1 Team", email = "contact@sonic-g1.dev"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["robotics", "reinforcement-learning", "motion-imitation", "humanoid", "unitree", "g1"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core ML/RL
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",

    # MuJoCo and simulation
    "mujoco>=2.3.0",
    "dm-control>=1.0.0",

    # Configuration and utilities
    "hydra-core>=1.2.0",
    "omegaconf>=2.2.0",
    "pyyaml>=6.0",

    # Data processing
    "pandas>=1.4.0",
    "h5py>=3.6.0",
    "pickle5>=0.0.11; python_version<'3.8'",

    # Visualization and plotting
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "opencv-python>=4.5.0",

    # Deployment and inference
    "onnx>=1.12.0",
    "onnxruntime>=1.12.0",

    # Logging and monitoring
    "wandb>=0.13.0",
    "tensorboard>=2.8.0",
    "tqdm>=4.64.0",

    # Utilities
    "click>=8.0.0",
    "rich>=12.0.0",
    "pathlib2>=2.3.6; python_version<'3.4'",
]

[project.optional-dependencies]
dev = [
    # Development tools
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",

    # Testing
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "pytest-mock>=3.7.0",

    # Documentation
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.17.0",

    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.9.0",
    "ipywidgets>=7.6.0",
]

gpu = [
    # GPU acceleration
    "torch>=1.12.0+cu116",
    "onnxruntime-gpu>=1.12.0",
]

tensorrt = [
    # TensorRT optimization
    "tensorrt>=8.4.0",
    "pycuda>=2022.1",
]

all = [
    "sonic-g1[dev,gpu,tensorrt]",
]

[project.urls]
Homepage = "https://github.com/sagarbpatel31/unitree_sonic_g1"
Repository = "https://github.com/sagarbpatel31/unitree_sonic_g1.git"
Documentation = "https://sonic-g1.readthedocs.io"
"Bug Tracker" = "https://github.com/sagarbpatel31/unitree_sonic_g1/issues"

[project.scripts]
sonic-g1-train = "sonic_g1.scripts.train:main"
sonic-g1-eval = "sonic_g1.scripts.eval:main"
sonic-g1-export = "sonic_g1.scripts.export:main"
sonic-g1-deploy = "sonic_g1.scripts.deploy:main"

[tool.setuptools]
packages = ["sonic_g1"]

[tool.setuptools.package-data]
sonic_g1 = [
    "assets/**/*",
    "configs/**/*.yaml",
    "configs/**/*.yml",
]

# Black code formatting
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort import sorting
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# MyPy type checking
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "mujoco.*",
    "dm_control.*",
    "onnxruntime.*",
    "tensorrt.*",
    "wandb.*",
    "cv2.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=sonic_g1",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
]

# Coverage configuration
[tool.coverage.run]
source = ["sonic_g1"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]